requirements.txt
pygame>=2.5.0
Pillow>=10.0.0
numpy<2.0.0
requests>=2.31.0
yfinance>=0.2.28
pandas>=2.0.0
matplotlib>=3.7.0
opencv-python>=4.8.0
PyYAML>=6.0.1
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.21.0
torch>=2.0.0
safetensors>=0.3.0
